Mac OS X gcc warns about a signed/unsigned comparison here. Explicit cast.
git-svn-id: svn://svn.tartarus.org/sgt/halibut@5188 cda61777-01e9-0310-a592-d414129be87e

+++ ustring.c
@@ -182,7 +182,7 @@ char *utoa_locale_dup(wchar_t const *s)
     siz = wcstombs(ret, s, len);
 
     if (siz) {
-	assert(siz <= MB_CUR_MAX * len);
+	assert(siz <= (size_t)(MB_CUR_MAX * len));
 	ret[siz] = '\0';
 	ret = sresize(ret, siz+1, char);
 	return ret;
